feat: validate driver with database engine (#1108)
Removing ambiguous behavior if a user configures a database
driver to connect to an incompatible database engine.

i.e. pymysql (MySQL driver) -> Cloud SQL for PostgreSQL database

This behavior would timeout trying to connect and then throw
an ambiguous timeout error.

The Connector.connect already know which database engine
the Cloud SQL database is prior to attempting to connect via the
driver. So we can validate if the driver is compatible and throw an
actionable error if not.