test: update server used for tests to use SSL/TLS (#1262)
This PR improves the local server used for unit tests.

The previous server fixture was just a plain socket unable to perform SSL/TLS.
This meant in certain tests we were skipping the SSL/TLS handshake because it
would fail.

The new proxy_server fixture more closely behaves to that of the real Proxy server.
It allows SSL/TLS by loading in the server CA used for testing, the client can then
use the new context fixture to properly use SSL to connect to the server.

This will provide a foundation for testing the metadata exchange which will be
introduced in the near future.