fix: error on connect_async() event loop mismatch (#1113)
Today, the Python Connector for async usage does not allow being called across
multiple event loops.

However, we do not explicitly define that logic nor make it clear via an error message.

This PR adds an error message to check in connect_async if the Connector._loop
attribute matches the current running event loop and errors if it does not (multiple event loops).

Related to #1107